Complete Vulkan overhaul and rework, Added support for Multithreading.

+++ Morpheus-Core/Source/Morpheus/Utilities/DeltaTime.h
@@ -7,18 +7,18 @@ namespace Morpheus {
 	class DeltaTime
 	{
 	public:
-		DeltaTime(float64 _Time = 0.00f)
+		DeltaTime(float32 _Time = 0.00f)
 			: m_Time(_Time)
 		{
 		}
 
-		float64 GetSeconds() const { return m_Time; }
-		float64 GetMilliseconds() const { return m_Time * 1000; }
+		float32 GetSeconds() const { return m_Time; }
+		float32 GetMilliseconds() const { return m_Time * 1000; }
 
-		operator float64() const { return m_Time; }
+		operator float32() const { return m_Time; }
 
 	private:
-		float64 m_Time;
+		float32 m_Time;
 
 	};
